# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CampusVoice API",
        "description": "API documentation for CampusVoice application",
        "contact": {
            "name": "CampusVoice Support",
            "email": "admin@campusvoice.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "CampusVoice API Server"
        }
    ],
    "paths": {
...
        "/suggestions": {
            "get": {
                "tags": [
                    "Suggestions"
                ],
                "summary": "Get all suggestions",
                "description": "Returns a list of suggestions based on user role. Admins and department heads see all suggestions, while regular users only see their own suggestions.",
                "operationId": "getSuggestions",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Suggestion"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to fetch suggestions"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Suggestions"
                ],
                "summary": "Create a new suggestion",
                "description": "Creates a new suggestion and returns it",
                "operationId": "storeSuggestion",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "department_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Extended Lab Hours"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "It would be beneficial to have the computer lab open for longer hours during exam periods."
                                    },
                                    "department_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Suggestion created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Suggestion created successfully"
                                        },
                                        "suggestion": {
                                            "$ref": "#/components/schemas/Suggestion"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to create suggestion"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Department": {
                "title": "Department",
                "description": "Department model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of the department",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name_ar": {
                        "title": "Name (Arabic)",
                        "description": "Name of the department in Arabic",
                        "type": "string",
                        "example": "قسم علوم الحاسوب"
                    },
                    "name_fr": {
                        "title": "Name (French)",
                        "description": "Name of the department in French",
                        "type": "string",
                        "example": "Département d'Informatique"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Creation date",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-05-20T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Last update date",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-05-20T12:00:00.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Department"
                }
            },
...
            "Suggestion": {
                "title": "Suggestion",
                "description": "Suggestion model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of the suggestion",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "Title",
                        "description": "Title of the suggestion",
                        "type": "string",
                        "example": "Extended Lab Hours"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Detailed description of the suggestion",
                        "type": "string",
                        "example": "It would be beneficial to have the computer lab open for longer hours during exam periods."
                    },
                    "status": {
                        "title": "Status",
                        "description": "Current status of the suggestion",
                        "type": "string",
                        "enum": [
                            "pending",
                            "approved",
                            "rejected"
                        ],
                        "example": "pending"
                    },
                    "user_id": {
                        "title": "User ID",
                        "description": "ID of the user who created the suggestion",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "department_id": {
                        "title": "Department ID",
                        "description": "ID of the department this suggestion is for",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "department": {
                        "$ref": "#/components/schemas/Department"
                    },
                    "vote_count": {
                        "title": "Vote Count",
                        "description": "Number of votes for this suggestion",
                        "type": "integer",
                        "example": 5
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Creation date",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-05-20T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Last update date",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-05-20T12:00:00.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Suggestion"
                }
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID of the user",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name of the user",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email address of the user",
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "role": {
                        "title": "Role",
                        "description": "Role of the user",
                        "type": "string",
                        "enum": [
                            "student",
                            "department_admin",
                            "admin"
                        ],
                        "example": "student"
                    },
                    "is_anonymous": {
                        "title": "Is Anonymous",
                        "description": "Whether the user is anonymous",
                        "type": "boolean",
                        "example": false
                    },
                    "department_id": {
                        "title": "Department ID",
                        "description": "ID of the department this user belongs to",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "department": {
                        "$ref": "#/components/schemas/Department"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Creation date",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-05-20T12:00:00.000000Z"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Last update date",
                        "type": "string",
                        "format": "datetime",
                        "example": "2025-05-20T12:00:00.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}